+++ src/test/groovy/org/codenarc/rule/groovyism/UseCollectNestedRuleTest.groovy
@@ -27,6 +27,18 @@ import org.junit.Test
 class UseCollectNestedRuleTest extends AbstractRuleTestCase<UseCollectNestedRule> {
 
     @Test
+    void test_InnerCollectResultTransformedIntoNonCollection_NoViolation() {
+        final SOURCE = '''
+            // The collect() call is just one part of a larger expression; the result may be transformed into a different type
+            [['a', 'b', 'c'], ['d', 'e']].collect { 'x' + it.collect { it + 'y' } }
+            
+            // The result from the collect() call is the object of another method call; the result may be a different type
+            [['a', 'b', 'c'], ['d', 'e']].collect { it.collect { it + 'y' }.join(' ') }
+        '''
+        assertNoViolations(SOURCE)
+    }
+
+    @Test
     void testRuleProperties() {
         assert rule.priority == 2
         assert rule.name == 'UseCollectNested'