+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Updated/Enhanced Rules and Bug Fixes
  - #651: **MissingOverrideAnnotation**: Ignore private superclass methods.
  - #649: Support for description inside CodeNarc enable/disable comments. ([davidkron](https://github.com/davidkron))
  - #652: **ClassStartsWithBlankLine**: Fix NPE on annotated class with no following line at end of source file.
+ - #653: **UseCollectNested**: Fix false positive when inner collect() call returns a different type.
 
 
 Version 2.2.0    (Aug 2021)